Added LSI model. Pulled LexicalBase out of TF-IDF to share components with LSI. Fixed bug in Working Set.
